feat(CampgroundForm): separate steps into files
